services/BlobStorageService/pyproject.toml
[project]
name = "blobstorageservice"
version = "0.1.0"
description = ""
authors = [
    {name = "yithril",email = "yithril@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob (>=12.26.0,<13.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

